requirements.txt
Flask>=2.3.0
pybit>=5.6.0
pyTelegramBotAPI>=4.20.0
python-dotenv>=1.0.1
requests>=2.31.0

